Made player script and transferred values to there
