+++ Assets/_Prefabs/Player.prefab
@@ -1,18 +1,5 @@
 %YAML 1.1
 %TAG !u! tag:unity3d.com,2011:
---- !u!114 &8411550192270314174
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 67184980228834723}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: ec284c942ab86044c9cd2251817c3d1b, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  LightObj: {fileID: 65604985700079799}
 --- !u!1 &4232917174339151221
 GameObject:
   m_ObjectHideFlags: 0
@@ -26,6 +13,8 @@ GameObject:
   - component: {fileID: 4232917174346240149}
   - component: {fileID: 4232917174344304149}
   - component: {fileID: 4232917174343056533}
+  - component: {fileID: 4431200693922674006}
+  - component: {fileID: 1084851102417055658}
   m_Layer: 0
   m_Name: Player
   m_TagString: Untagged
@@ -302,6 +291,32 @@ AudioSource:
     m_PreInfinity: 2
     m_PostInfinity: 2
     m_RotationOrder: 0
+--- !u!114 &4431200693922674006
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 4232917174339151221}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: ec284c942ab86044c9cd2251817c3d1b, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  LightObj: {fileID: 65604985700079799}
+--- !u!114 &1084851102417055658
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 4232917174339151221}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 4ff83bb35ea252749b5c40b6de34b4e2, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  _batteryDrainRate: 0.1
 --- !u!1 &4232917174339151223
 GameObject:
   m_ObjectHideFlags: 0
@@ -472,12 +487,6 @@ PrefabInstance:
       objectReference: {fileID: 0}
     m_RemovedComponents: []
   m_SourcePrefab: {fileID: 100100000, guid: d86ca2f4159e6a942abd8dfc52b38395, type: 3}
---- !u!1 &67184980228834723 stripped
-GameObject:
-  m_CorrespondingSourceObject: {fileID: 1002705123656268, guid: d86ca2f4159e6a942abd8dfc52b38395,
-    type: 3}
-  m_PrefabInstance: {fileID: 66779484207826927}
-  m_PrefabAsset: {fileID: 0}
 --- !u!4 &71431574375225147 stripped
 Transform:
   m_CorrespondingSourceObject: {fileID: 4777606630871252, guid: d86ca2f4159e6a942abd8dfc52b38395,
