+++ Assets/_Scripts/Player.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 4ff83bb35ea252749b5c40b6de34b4e2
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
